Split Arc and Sector modules into more files
