+++ embedded-graphics/src/primitives/arc/mod.rs
@@ -1,0 +1,229 @@
+//! The arc primitive
+
+mod linear_equation;
+mod plane_sector;
+mod points;
+mod styled;
+
+use crate::{
+    geometry::{Angle, Dimensions, Point, Size},
+    primitives::{Primitive, Rectangle},
+    transform::Transform,
+};
+pub(in crate::primitives) use linear_equation::LinearEquation;
+pub(in crate::primitives) use plane_sector::{PlaneSector, PlaneSectorIterator};
+pub use points::Points;
+pub use styled::StyledPixels;
+
+/// Arc primitive
+///
+/// # Examples
+///
+/// ## Create some arcs with different styles
+///
+/// ```rust
+/// use embedded_graphics::{
+///     pixelcolor::Rgb565,
+///     prelude::*,
+///     primitives::Arc,
+///     style::{PrimitiveStyle, PrimitiveStyleBuilder},
+/// };
+/// # use embedded_graphics::mock_display::MockDisplay;
+/// # let mut display = MockDisplay::default();
+///
+/// // Arc with 1 pixel wide white stroke with top-left point at (10, 20) with a diameter of 30
+/// Arc::new(Point::new(10, 20), 30, 0.0.deg(), 90.0.deg())
+///     .into_styled(PrimitiveStyle::with_stroke(Rgb565::WHITE, 1))
+///     .draw(&mut display)?;
+///
+/// // Arc with styled stroke with top-left point at (15, 25) with a diameter of 20
+/// let style = PrimitiveStyleBuilder::new()
+///     .stroke_color(Rgb565::RED)
+///     .stroke_width(3)
+///     .build();
+///
+/// Arc::new(Point::new(15, 25), 20, 180.0.deg(), -90.0.deg())
+///     .into_styled(style)
+///     .draw(&mut display)?;
+/// # Ok::<(), core::convert::Infallible>(())
+/// ```
+#[derive(Copy, Clone, PartialEq, PartialOrd, Debug)]
+pub struct Arc {
+    /// Top-left point of the bounding-box of the circle supporting the arc
+    pub top_left: Point,
+
+    /// Diameter of the circle supporting the arc
+    pub diameter: u32,
+
+    /// Angle at which the arc starts
+    pub angle_start: Angle,
+
+    /// Angle defining the arc sweep starting at angle_start
+    pub angle_sweep: Angle,
+}
+
+impl Arc {
+    /// Create a new arc delimited with a top-left point with a specific diameter and start and sweep angles
+    pub const fn new(
+        top_left: Point,
+        diameter: u32,
+        angle_start: Angle,
+        angle_sweep: Angle,
+    ) -> Self {
+        Arc {
+            top_left,
+            diameter,
+            angle_start,
+            angle_sweep,
+        }
+    }
+
+    /// Create a new arc centered around a given point with a specific diameter and start and sweep angles
+    pub fn with_center(
+        center: Point,
+        diameter: u32,
+        angle_start: Angle,
+        angle_sweep: Angle,
+    ) -> Self {
+        let top_left = center - Size::new_equal(diameter).center_offset();
+
+        Arc {
+            top_left,
+            diameter,
+            angle_start,
+            angle_sweep,
+        }
+    }
+
+    /// Return the center point of the arc
+    pub fn center(&self) -> Point {
+        self.bounding_box().center()
+    }
+
+    /// Return the center point of the arc scaled by a factor of 2
+    ///
+    /// This method is used to accurately calculate the outside edge of the arc.
+    /// The result is not equivalent to `self.center() * 2` because of rounding.
+    fn center_2x(&self) -> Point {
+        // The radius scaled up by a factor of 2 is equal to the diameter
+        let radius = self.diameter.saturating_sub(1);
+
+        self.top_left * 2 + Size::new(radius, radius)
+    }
+
+    pub(crate) fn expand(&self, offset: u32) -> Self {
+        let diameter = self.diameter.saturating_add(2 * offset);
+
+        Self::with_center(self.center(), diameter, self.angle_start, self.angle_sweep)
+    }
+
+    pub(crate) fn shrink(&self, offset: u32) -> Self {
+        let diameter = self.diameter.saturating_sub(2 * offset);
+
+        Self::with_center(self.center(), diameter, self.angle_start, self.angle_sweep)
+    }
+}
+
+impl Primitive for Arc {
+    type PointsIter = Points;
+
+    fn points(&self) -> Self::PointsIter {
+        Points::new(self)
+    }
+}
+
+impl Dimensions for Arc {
+    fn bounding_box(&self) -> Rectangle {
+        Rectangle::new(self.top_left, Size::new(self.diameter, self.diameter))
+    }
+}
+
+impl Transform for Arc {
+    /// Translate the arc from its current position to a new position by (x, y) pixels,
+    /// returning a new `Arc`. For a mutating transform, see `translate_mut`.
+    ///
+    /// ```
+    /// # use embedded_graphics::primitives::Arc;
+    /// # use embedded_graphics::prelude::*;
+    /// let arc = Arc::new(Point::new(5, 10), 10, 0.0.deg(), 90.0.deg());
+    /// let moved = arc.translate(Point::new(10, 10));
+    ///
+    /// assert_eq!(moved.top_left, Point::new(15, 20));
+    /// ```
+    fn translate(&self, by: Point) -> Self {
+        Self {
+            top_left: self.top_left + by,
+            ..*self
+        }
+    }
+
+    /// Translate the arc from its current position to a new position by (x, y) pixels.
+    ///
+    /// ```
+    /// # use embedded_graphics::primitives::Arc;
+    /// # use embedded_graphics::prelude::*;
+    /// let mut arc = Arc::new(Point::new(5, 10), 10, 0.0.deg(), 90.0.deg());
+    /// arc.translate_mut(Point::new(10, 10));
+    ///
+    /// assert_eq!(arc.top_left, Point::new(15, 20));
+    /// ```
+    fn translate_mut(&mut self, by: Point) -> &mut Self {
+        self.top_left += by;
+
+        self
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::geometry::AngleUnit;
+
+    #[test]
+    fn negative_dimensions() {
+        let arc = Arc::new(Point::new(-15, -15), 10, 0.0.deg(), 90.0.deg());
+
+        assert_eq!(
+            arc.bounding_box(),
+            Rectangle::new(Point::new(-15, -15), Size::new(10, 10))
+        );
+    }
+
+    #[test]
+    fn dimensions() {
+        let arc = Arc::new(Point::new(5, 15), 10, 0.0.deg(), 90.0.deg());
+
+        assert_eq!(
+            arc.bounding_box(),
+            Rectangle::new(Point::new(5, 15), Size::new(10, 10))
+        );
+    }
+
+    #[test]
+    fn it_handles_negative_coordinates() {
+        let positive = Arc::new(Point::new(10, 10), 5, 0.0.deg(), 90.0.deg()).points();
+
+        let negative = Arc::new(Point::new(-10, -10), 5, 0.0.deg(), 90.0.deg()).points();
+
+        assert!(negative.eq(positive.map(|p| p - Point::new(20, 20))));
+    }
+
+    #[test]
+    fn center_is_correct() {
+        // odd diameter
+        let arc = Arc::new(Point::new(10, 10), 5, 0.0.deg(), 90.0.deg());
+        assert_eq!(arc.center(), Point::new(12, 12));
+
+        // even diameter
+        let arc = Arc::new(Point::new(10, 10), 6, 0.0.deg(), 90.0.deg());
+        assert_eq!(arc.center(), Point::new(12, 12));
+
+        // odd diameter
+        let arc = Arc::with_center(Point::new(10, 10), 5, 0.0.deg(), 90.0.deg());
+        assert_eq!(arc.center(), Point::new(10, 10));
+
+        // even diameter
+        let arc = Arc::with_center(Point::new(10, 10), 6, 0.0.deg(), 90.0.deg());
+        assert_eq!(arc.center(), Point::new(10, 10));
+    }
+}
