+++ embedded-graphics/src/primitives/arc/styled.rs
@@ -1,0 +1,171 @@
+use crate::{
+    drawable::{Drawable, Pixel},
+    geometry::Point,
+    pixelcolor::PixelColor,
+    primitives::{
+        arc::{plane_sector::PlaneSectorIterator, Arc},
+        circle,
+        circle::DistanceIterator,
+        Styled,
+    },
+    style::PrimitiveStyle,
+    DrawTarget,
+};
+
+/// Pixel iterator for each pixel in the arc border
+#[derive(Copy, Clone, PartialEq, PartialOrd, Debug)]
+pub struct StyledPixels<C>
+where
+    C: PixelColor,
+{
+    iter: DistanceIterator<PlaneSectorIterator>,
+
+    outer_threshold: u32,
+    inner_threshold: u32,
+
+    stroke_color: Option<C>,
+}
+
+impl<C> StyledPixels<C>
+where
+    C: PixelColor,
+{
+    fn new(styled: &Styled<Arc, PrimitiveStyle<C>>) -> Self {
+        let Styled { primitive, style } = styled;
+
+        let stroke_area = primitive.expand(style.outside_stroke_width());
+        let fill_area = primitive.shrink(style.inside_stroke_width());
+
+        let inner_threshold = circle::diameter_to_threshold(fill_area.diameter);
+        let outer_threshold = circle::diameter_to_threshold(stroke_area.diameter);
+
+        let iter = if !styled.style.is_transparent() {
+            DistanceIterator::new(
+                stroke_area.center_2x(),
+                PlaneSectorIterator::new(
+                    &stroke_area,
+                    stroke_area.center(),
+                    stroke_area.angle_start,
+                    stroke_area.angle_sweep,
+                ),
+            )
+        } else {
+            DistanceIterator::new(Point::zero(), PlaneSectorIterator::empty())
+        };
+
+        Self {
+            iter,
+            outer_threshold,
+            inner_threshold,
+            stroke_color: styled.style.stroke_color,
+        }
+    }
+}
+
+impl<C> Iterator for StyledPixels<C>
+where
+    C: PixelColor,
+{
+    type Item = Pixel<C>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let stroke_color = self.stroke_color?;
+        let outer_threshold = self.outer_threshold;
+        let inner_threshold = self.inner_threshold;
+
+        self.iter
+            .find(|(_, distance)| *distance < outer_threshold && *distance >= inner_threshold)
+            .map(|(point, _)| Pixel(point, stroke_color))
+    }
+}
+
+impl<'a, C: 'a> Drawable<C> for &Styled<Arc, PrimitiveStyle<C>>
+where
+    C: PixelColor,
+{
+    fn draw<D: DrawTarget<Color = C>>(self, display: &mut D) -> Result<(), D::Error> {
+        display.draw_iter(self)
+    }
+}
+
+impl<'a, C> IntoIterator for &'a Styled<Arc, PrimitiveStyle<C>>
+where
+    C: PixelColor,
+{
+    type Item = Pixel<C>;
+    type IntoIter = StyledPixels<C>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        StyledPixels::new(self)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{
+        geometry::AngleUnit,
+        mock_display::MockDisplay,
+        pixelcolor::BinaryColor,
+        primitives::Primitive,
+        style::{PrimitiveStyle, PrimitiveStyleBuilder, StrokeAlignment},
+    };
+
+    // Check the rendering of a simple arc
+    #[test]
+    fn tiny_arc() -> Result<(), core::convert::Infallible> {
+        let mut display = MockDisplay::new();
+
+        Arc::new(Point::zero(), 7, 30.0.deg(), 120.0.deg())
+            .into_styled(PrimitiveStyle::with_stroke(BinaryColor::On, 1))
+            .draw(&mut display)?;
+
+        #[rustfmt::skip]
+        assert_eq!(
+            display,
+            MockDisplay::from_pattern(&[
+                "  ###  ",
+                " #   # ",
+            ])
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn stroke_alignment() {
+        const CENTER: Point = Point::new(15, 15);
+        const SIZE: u32 = 10;
+
+        let style = PrimitiveStyle::with_stroke(BinaryColor::On, 3);
+
+        let mut display_center = MockDisplay::new();
+        Arc::with_center(CENTER, SIZE, 0.0.deg(), 90.0.deg())
+            .into_styled(style)
+            .draw(&mut display_center)
+            .unwrap();
+
+        let mut display_inside = MockDisplay::new();
+        Arc::with_center(CENTER, SIZE + 2, 0.0.deg(), 90.0.deg())
+            .into_styled(
+                PrimitiveStyleBuilder::from(&style)
+                    .stroke_alignment(StrokeAlignment::Inside)
+                    .build(),
+            )
+            .draw(&mut display_inside)
+            .unwrap();
+
+        let mut display_outside = MockDisplay::new();
+        Arc::with_center(CENTER, SIZE - 4, 0.0.deg(), 90.0.deg())
+            .into_styled(
+                PrimitiveStyleBuilder::from(&style)
+                    .stroke_alignment(StrokeAlignment::Outside)
+                    .build(),
+            )
+            .draw(&mut display_outside)
+            .unwrap();
+
+        assert_eq!(display_center, display_inside);
+        assert_eq!(display_center, display_outside);
+    }
+}
